deal with bug result type Float can't be cast to the desired output type __int64

+++ lib/core/postprocess.py
@@ -27,7 +27,7 @@ def build_targets(hyp, predictions, targets, model):
     # print(type(det))
     na, nt = det.na, targets.shape[0]  # number of anchors, targets
     tcls, tbox, indices, anch = [], [], [], []
-    gain = torch.ones(7, device=targets.device,dtype=torch.int64)  # normalized to gridspace gain
+    gain = torch.ones(7, device=targets.device)  # normalized to gridspace gain
     ai = torch.arange(na, device=targets.device).float().view(na, 1).repeat(1, nt)  # same as .repeat_interleave(nt)
     targets = torch.cat((targets.repeat(na, 1, 1), ai[:, :, None]), 2)  # append anchor indices
     
@@ -71,7 +71,7 @@ def build_targets(hyp, predictions, targets, model):
         # Append
         a = t[:, 6].long()  # anchor indices
 
-        indices.append((b, a, gj.clamp_(0, gain[3] - 1), gi.clamp_(0, gain[2] - 1)))  # image, anchor, grid indices
+        indices.append((b, a, gj.clamp_(0, gain[3].type(torch.int64) - 1), gi.clamp_(0, gain[2].type(torch.int64) - 1)))  # image, anchor, grid indices
         tbox.append(torch.cat((gxy - gij, gwh), 1))  # box
         anch.append(anchors[a])  # anchors
         tcls.append(c)  # class
